Create hosting_101.js
	Hoisting:
a.	Variables being hoisted to the top of the scope (JS way of making space for upcoming variables.) A variable that is not yet defined is still hoisted to the top.
b.	If that hoisted variable is defined with var there is no reference error, an undefined is given
c.	If a variable is defined with const and let -> it goes into the temporal dead zone, thus is not hoisted and will give a reference error if we try to log it before initialization